Recreate the Mandrill Clients for each test
workaround horribly unreliable Akka HTTP Pool
